Network Homework/library_management/app/helpers/utils.py
from jose import jwt, JWTError
from app.helpers.config import settings
from fastapi import HTTPException, status
from passlib.context import CryptContext
from datetime import datetime, timedelta , timezone

pwd_context = CryptContext(schemes=["bcrypt"])

def verify_password(plain_password: str, hashed_password: str):
  return pwd_context.verify(plain_password, hashed_password)

def hash_password(password: str)->str:
  return pwd_context.hash(password)

def create_jwt(data: dict):
  if not isinstance(data, dict):
        data = data.dict()
  to_encode = data.copy()
  expire = datetime.now(timezone.utc) + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
  to_encode.update({"exp": expire})
  return jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)

def decode_jwt(token: str):
  try:
    payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
    return payload
  except JWTError:
    raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
  